+++ mdsf/src/tools/yq.rs
@@ -16,3 +16,34 @@ pub fn set_args(
 pub const COMMANDS: [CommandType; 1] = [CommandType::Direct("yq")];
 
 pub const IS_STDIN: bool = false;
+
+#[cfg(test)]
+mod test_yq {
+    #[test_with::executable(yq)]
+    fn test_yq_json_b20bccf3f90b7945() {
+        let input = r#"{ "yq": "yq"  }"#;
+
+        let output = r#"{
+  "yq": "yq"
+}
+"#;
+
+        let file_ext = crate::fttype::get_file_extension("json");
+
+        let snippet =
+            crate::execution::setup_snippet(input, &file_ext).expect("it to create a snippet file");
+
+        let result = crate::tools::Tooling::Yq
+            .format_snippet(
+                snippet.path(),
+                crate::testing::DEFAULT_TEST_FORMATTER_TIMEOUT,
+                crate::testing::DEFAULT_TEST_DEBUG_ENABLED,
+                &crate::config::MdsfConfigRunners::all(),
+            )
+            .expect("it to be successful")
+            .1
+            .expect("it to be some");
+
+        assert_eq!(result, output);
+    }
+}